Move string-based argument conversion tests to `platform-test`
Prior to this commit, `platform-tests` had no dedicated tests for
`ConversionSupport`. Instead, `DefaultArgumentConverterTests` from
`jupiter-tests` provided the corresponding coverage.

Now, string-based tests are moved to `ConversionSupportTests`, while
`DefaultArgumentConverterTests` verify that the delegation to
`ConversionSupport` happens correctly.
